fix: move task 2 core test coverage
